+++ NEWS.md
@@ -1,4 +1,4 @@
-ctmm 0.6.2 (2021-11-03)
+ctmm 0.6.2 (2021-12-12)
 ================
 
   * new function rsf.fit() to fit integrated resource selection functions (RSFs) with autocorrelation-adjusted weighted likelihood
@@ -6,10 +6,12 @@ ctmm 0.6.2 (2021-11-03)
   * new function revisitation() to calculate the distribution of revisitations.
   
   * new function npr() to calculate non-parametric spatial regressions
-
+  
   * new argument SP in plot.telemetry and plot.UD() for plotting shapefile base layers
   
-  * bugfix in ctmm.loglike() when fitting multiple error classes, where zome are zero
+  * tbind now works correctly with different projections and calibrations
+  
+  * bugfix in ctmm.loglike() when fitting multiple error classes, where some are zero
 
 ctmm 0.6.1 (2021-07-26)
 ================